Move RPC relay to a dedicated EthForwarder module
